refactor(ingest/s3): move DuckDB profiler to data_lake_common (shared by data-lake sources)
Addresses review feedback: the DuckDB profiler is generic data-lake profiling
(S3 + GCS-over-S3, and the intended shared engine), so it does not belong under
source/s3/. Move duckdb_profiler.py and duckdb_secrets.py to
source/data_lake_common/ (the existing shared home for path_spec/data_lake_utils
used by s3/gcs/abs/delta) and relocate their unit tests to mirror.

To keep data_lake_common from depending on a source package, the profiler now
types its report against a small structural Protocol (report_warning +
report_file_dropped) instead of importing s3's DataLakeSourceReport; the
concrete source reports satisfy it. TableData stays a type-only import.

Pure relocation — no behavior change. s3/source.py imports the profiler from
the new location.
