+++ metadata-ingestion/src/datahub/ingestion/source/data_lake_common/duckdb_secrets.py
